fix: [GTM-712]: Use default signup notification template OnPrem (#32273)
* [GTM-712]: Use default signup notification template OnPrem
